requirements.txt
flask==3.1.0
langchain-chroma==0.2.2
langchain-core==0.3.45
langchain-openai==0.3.7
pandas==2.2.3
python-dotenv==1.0.1
